+++ services/post/src/modules/command/post/post.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { QueryController } from 'src/modules/query/query.controller';
+import { QueryController } from 'src/modules/query/comment/query.controller';
 import { PostController } from './post.controller';
 import { PostService } from './post.service';
