--- services/post/src/modules/query/post.query.dto.ts
@@ -1,10 +1,0 @@
-export interface PostQueryDto {
-  id: number;
-  author: number;
-  createdAt: Date;
-  updatedAt: Date;
-  slug: string;
-  images: string[];
-  content: string;
-  likenum: number;
-}
